+++ packages/esix/src/index.ts
@@ -2,7 +2,12 @@ import BaseModel from './base-model'
 import { ConnectionHandler, connectionHandler } from './connection-handler'
 import { getCollectionName } from './naming'
 import QueryBuilder, { type Query } from './query-builder'
-import { type ObjectType, type Dictionary, type Document } from './types'
+import {
+  type ObjectType,
+  type Dictionary,
+  type Document,
+  type Paginated
+} from './types'
 
 export {
   BaseModel,
@@ -11,4 +16,4 @@ export {
   connectionHandler,
   getCollectionName
 }
-export type { Query, ObjectType, Dictionary, Document }
+export type { Query, ObjectType, Dictionary, Document, Paginated }